Extract guest user logic into a Concern.
