+++ app/controllers/concerns/guest_user.rb
@@ -1,0 +1,16 @@
+module GuestUser
+  extend ActiveSupport::Concern
+
+  private
+
+    def create_guest_user
+      user = User.new(email: "#{SecureRandom.uuid}@#{SecureRandom.uuid}.com", accepted_terms: true, guest: true)
+      user.save!(validate: false)
+      session[:guest_user_id] = user.id
+      user
+    end 
+
+    def set_guest_user
+      @user = User.find_by(id: session[:guest_user_id]).present? ? User.find_by(id: session[:guest_user_id]) : create_guest_user
+    end
+end
